Threat Head
+ Basic Threat: A floating head hehe
+ Threats are like peeps, they can have tasks
+ Walk task can now successfully perform offroad (needs to limit water tiles, etc.)
+ Fixed tasks starting with wait, now waits only one frame before starting
+ Basic threat just moves towards tile hero is on

+++ AemonsNookMono/Entities/Entity.cs
@@ -9,8 +9,8 @@ namespace AemonsNookMono.Entities
     public class Entity
     {
         #region Public Properties
-        public int CenterX { get; set; }
-        public int CenterY { get; set; }
+        public float CenterX { get; set; }
+        public float CenterY { get; set; }
         public Tile TileOn { get; set; }
         public Tile EntranceTile { get; set; }
         public Tile ExitTile { get; set; }
